+++ core/src/geometry/shapes.rs
@@ -43,7 +43,7 @@ pub const CUBE_TRIS: [usize; 36] = [
 pub struct Cube(pub Object);
 
 impl Cube {
-    pub fn new(size: f64) -> Self {
+    pub fn new(size: f32) -> Self {
         let c_mesh = Mesh::from_vertices_faces(CUBE_VERTS.into(), CUBE_TRIS.into());
 
         let transform = Transform {
